fix(gateway,fault_manager): fix format specifier portability issues
Use PRId64 with static_cast<int64_t> for portable printf formatting
of parameter values and chrono durations. The underlying type of
declare_parameter<int> and chrono::milliseconds::rep varies across
platforms (long vs long long), so explicit casts ensure format
specifier agreement.

+++ src/ros2_medkit_fault_manager/src/fault_manager_node.cpp
@@ -17,6 +17,7 @@
 #include <yaml-cpp/yaml.h>
 
 #include <cctype>
+#include <cinttypes>
 #include <filesystem>
 #include <fstream>
 #include <nlohmann/json.hpp>
@@ -110,7 +111,8 @@ FaultManagerNode::FaultManagerNode(const rclcpp::NodeOptions & options) : Node("
   }
   auto max_snapshots = declare_parameter<int>("snapshots.max_per_fault", 10);
   if (max_snapshots < 0) {
-    RCLCPP_WARN(get_logger(), "snapshots.max_per_fault should be >= 0, got %ld. Disabling limit.", max_snapshots);
+    RCLCPP_WARN(get_logger(), "snapshots.max_per_fault should be >= 0, got %" PRId64 ". Disabling limit.",
+                static_cast<int64_t>(max_snapshots));
     max_snapshots = 0;
   }
